requirements.txt
pydantic
requests
uvicorn
pymongo
gunicorn
fastapi
sqlalchemy
python-jose
passlib
PyJWT